[ADD] Added bulk delete management

+++ components/data-table.tsx
@@ -22,7 +22,7 @@ import {
   TableHeader,
   TableRow,
 } from '@/components/ui/table'
-
+import { useConfirm } from '@/hooks/use-confirm'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import React from 'react'
@@ -43,6 +43,10 @@ export function DataTable<TData, TValue>({
   onDelete,
   disabled,
 }: DataTableProps<TData, TValue>) {
+  const [ConfirmDialog, confirm] = useConfirm(
+    'Are you sure?',
+    'You are about to perfom a bulk delete.'
+  )
   const [sorting, setSorting] = React.useState<SortingState>([])
   const [columnFilters, setColumnFilters] = React.useState<ColumnFiltersState>(
     []
@@ -69,6 +73,7 @@ export function DataTable<TData, TValue>({
 
   return (
     <div>
+      <ConfirmDialog />
       <div className='flex items-center py-4'>
         <Input
           placeholder={`Filter ${filterKey}...`}
@@ -84,9 +89,15 @@ export function DataTable<TData, TValue>({
             size='sm'
             variant='outline'
             className='ml-auto text-xs font-normal'
-            onClick={() => {
-              onDelete(table.getFilteredSelectedRowModel().rows)
-              table.resetRowSelection
+            onClick={async () => {
+              const ok = await confirm()
+              //Promesa utilizando hook propio "use-confirm"
+              //si ok = true, delete si no, no pasa nada
+              //muestra Dialog de shadcn
+              if (ok) {
+                onDelete(table.getFilteredSelectedRowModel().rows)
+                table.resetRowSelection
+              }
             }}
           >
             <Trash className='mr-2 size-4' />
